+++ server/src/test/java/org/opensearch/indices/IndicesServiceCloseTests.java
@@ -37,7 +37,8 @@ import org.apache.lucene.search.Query;
 import org.opensearch.cluster.ClusterName;
 import org.opensearch.cluster.routing.allocation.DiskThresholdSettings;
 import org.opensearch.common.cache.RemovalNotification;
-import org.opensearch.common.cache.tier.TierType;
+import org.opensearch.common.cache.tier.CacheValue;
+import org.opensearch.common.cache.tier.enums.CacheStoreType;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.common.util.concurrent.OpenSearchExecutors;
 import org.opensearch.core.common.bytes.BytesArray;
@@ -329,7 +330,7 @@ public class IndicesServiceCloseTests extends OpenSearchTestCase {
             }
 
             @Override
-            public void onCached(Key key, BytesReference value, TierType tierType) {}
+            public void onCached(Key key, BytesReference value, CacheStoreType tierType) {}
 
             @Override
             public boolean isOpen() {